chore(refactor): move jira code around
